BEN-1073 This commit does two major things:
* Properly hands the extra host compiler flags to cuda, so cuda
  can forward them to the host compiler for compilation of host
  code in .cu files.  Not doing this causes inconsistencies in
  the building of host code, which at the least causes us not to
  emit our usual level of strict warnings, and at the worst could
  cause ODR violations of any structure definitions were keying
  off compiler flags (like our m512 classes)
* Fully revived the intel build.
  * As a minor side effect of this we now enable the preprocessor
    definition used to work around broken constexprs in some cases.
    This seems to originally be put in place to work around deficiencies
    in some versions of the microsoft compiler, but nvcc appears to be
    broken in the same way and this preprocessor definition allows us to
    sidestep some bugs in nvcc
  * As a minor side effect of this, usages of <random> have been pushed
    to .cpp files only.  Enabling SSE3 or higher causes it to use some
    intrinsics that trigger another bug in nvcc when processing host
    code in .cu files.

The rest of the commit is just dealing with various warning messages that
are specific to either intel or .cu files, and thus were not seen previously

+++ src/basecaller/basecaller/test/TestTraceHistogramAccumHost.cpp
@@ -119,7 +119,7 @@ TEST_F(TestTraceHistogramAccumHost, DISABLED_WIP_One)
 
     // TODO: Blocked by incompleteness of BaselineStats. See BEN-896.
 
-    auto baselinedTracesAndStats = GenerateCamTraceBatch(blMean);
+    auto baselinedTracesAndStats = GenerateCamTraceBatch(static_cast<uint16_t>(blMean));
     tha.AddBatch(baselinedTracesAndStats.first, baselinedTracesAndStats.second.baselinerStats);
     EXPECT_EQ(chunkSize, tha.FramesAdded());
     EXPECT_EQ(0, tha.HistogramFrameCount());
